pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "py-prompt-versioner"
version = "1.0.1"
authors = [
  { name="Darryl See Wei Shen", email="dsee007@yahoo.com" },
]
description = "Lightweight prompt version control."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "python-frontmatter>=1.1.0",
]

[project.urls]
"Homepage" = "https://github.com/peanutsee/py-prompt-versioner"
"Bug Tracker" = "https://github.com/peanutsee/py-prompt-versioner/issues"

[project.scripts]
prompt-versioner = "py_prompt_versioner.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[tool.pytest.ini_options]
addopts = "--cov=py_prompt_versioner --cov-report=xml --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
source = ["py_prompt_versioner"]
branch = true

[tool.coverage.report]
show_missing = true
